requirements.txt
asgiref==3.4.1
Django==3.2.7
django-ckeditor==6.1.0
django-jazzmin==2.4.8
django-js-asset==1.2.2
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.1
